PROCESS : Reused metadata parser in tests.
Refactored the metadata parsing function used in test files and moved it to support file for better reusability.
