+++ frontend/tests/outcome.js
@@ -1,4 +1,5 @@
 import { Then as Alors, Step } from '@badeball/cypress-cucumber-preprocessor';
+import { parseStrToObject } from './support';
 
 Alors("la glose ouverte a le titre par défaut", () => {
   Step(this, "'…' est la glose ouverte");
@@ -140,17 +141,6 @@ Alors("la glose ouverte a les métadonnées", (metadata) => {
   cy.get('.icon.edit').click();
   cy.get('.editable.metadata').click();
   cy.get('form textarea').invoke('val').then(actual => {
-    const parseStrToObject = str => {
-      const lines = str.trim().split('\n');
-      const result = {};
-      lines.forEach(line => {
-        const [key, value] = line.split(':').map(part => part.trim());
-        if (key !== undefined && value !== undefined) {
-          result[key] = value;
-        }
-      });
-      return result;
-    };
     const expectedMetadata = parseStrToObject(metadata);
     const actualMetadata = parseStrToObject(actual);
     Object.entries(expectedMetadata).forEach(([key, value]) => {
